gen-ai/Assistants/notebooks/requirements.txt
openai==1.30.1
pyautogen==0.2.27
python-dotenv
yfinance
pillow
